requirements.txt
streamlit~=1.35.0
dateparser
sqlalchemy
millify
openai
python-dotenv
